Move httptransport templates to subpackage
