+++ gengokit/httptransport/templates_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/TuneLab/go-truss/gengokit/gentesthelper"
+	"github.com/TuneLab/go-truss/gengokit/httptransport/templates"
 )
 
 // Test that rendering certain templates will ouput the code we expect. The
@@ -215,7 +216,7 @@ func DecodeHTTPSumZeroRequest(_ context.Context, r *http.Request) (interface{}, 
 // Test that all the templated source code is identical to the source code
 // found within the file 'embeddable_funcs.go'.
 func TestHTTPAssistFuncs(t *testing.T) {
-	tmplfncs := FormatCode(HTTPAssistFuncs)
+	tmplfncs := FormatCode(templates.HTTPAssistFuncs)
 	// Get the source code for all the functions in the same source file as
 	// the BuildParamMap function
 	source, err := AllFuncSourceCode(BuildParamMap)
